Move maze logic to shared JS module
