+++ pc-gyro.html
@@ -5,6 +5,7 @@
     <title>PC - ジャイロ迷路</title>
     <link rel="stylesheet" href="css/styles.css" />
     <script src="https://cdn.socket.io/4.7.5/socket.io.min.js"></script>
+    <script src="js/maze-logic.js" defer></script>
     <script src="js/navigation-endpoint.js" defer></script>
     <script src="js/pc-gyro.js" defer></script>
   </head>
