+++ pc-next.html
@@ -5,6 +5,7 @@
     <title>PC - 迷路</title>
     <link rel="stylesheet" href="css/styles.css" />
     <script src="https://cdn.socket.io/4.7.5/socket.io.min.js"></script>
+    <script src="js/maze-logic.js" defer></script>
     <script src="js/navigation-endpoint.js" defer></script>
     <script src="js/pc-next.js" defer></script>
   </head>
